refactor: ScanWorkRepository 중복 쿼리 통합 및 ScanPersister 공통 로직 추출
